string to int fix

+++ modules/cogs/pinground.py
@@ -14,7 +14,7 @@ class PingRound(commands.Cog):
         if not (ctx.author.id == KAPER_DISCORD_ID or ctx.author.id == HELTROSH_DISCORD_ID):
             await ctx.send(ADMIN_ONLY_ERROR)
             return
-        elif not (round.isnumeric() and league.isnumeric() and (1 <= league <= 6) and (1 <= round <= 9)):
+        elif not (round.isnumeric() and league.isnumeric() and (1 <= int(league) <= 6) and (1 <= int(round) <= 9)):
             await ctx.send(WRONG_ROUND_OR_LEAGUE)
             return
         i=0
